Move loggerSet into its own file
